Move EnvoyQuicClock out of quic/platform
