+++ source/common/quic/envoy_quic_clock.cc
@@ -1,4 +1,4 @@
-#include "source/common/quic/platform/envoy_quic_clock.h"
+#include "source/common/quic/envoy_quic_clock.h"
 
 namespace Envoy {
 namespace Quic {
