+++ test/common/quic/envoy_quic_clock_test.cc
@@ -1,6 +1,6 @@
 #include <memory>
 
-#include "source/common/quic/platform/envoy_quic_clock.h"
+#include "source/common/quic/envoy_quic_clock.h"
 
 #include "test/test_common/simulated_time_system.h"
 #include "test/test_common/test_time.h"
