+++ test/common/quic/platform/BUILD
@@ -165,15 +165,3 @@ envoy_cc_test_library(
         ":quic_platform_expect_bug_impl_lib",
     ],
 )
-
-envoy_cc_test(
-    name = "envoy_quic_clock_test",
-    srcs = ["envoy_quic_clock_test.cc"],
-    tags = ["nofips"],
-    deps = [
-        "//source/common/quic/platform:envoy_quic_clock_lib",
-        "//test/test_common:simulated_time_system_lib",
-        "//test/test_common:test_time_lib",
-        "//test/test_common:utility_lib",
-    ],
-)
